+++ lib/glimmer/dsl/widget_expression.rb
@@ -6,17 +6,17 @@ module Glimmer
   module DSL
     class WidgetExpression < Expression
       include ParentExpression
-      
+
       EXCLUDED_KEYWORDS = %w[shell display tab_item]
 
       def can_interpret?(parent, keyword, *args, &block)
         !EXCLUDED_KEYWORDS.include?(keyword) and
-          widget?(parent) and
+          widget?(parent) and #TODO change to composite?(parent)
           SWT::WidgetProxy.widget_exists?(keyword)
       end
 
       def interpret(parent, keyword, *args, &block)
-        Glimmer.logger&.debug "widget styles are: " + args.inspect
+        Glimmer.logger&.debug "widget styles are: #{args}"
         SWT::WidgetProxy.new(keyword, parent, args)
       end
     end